// angular-oidc-ngrx/src/app/core/auth/oidc-event.enum.ts
export enum OidcEvent {
  AccessTokenExpired = 'addAccessTokenExpired',
  AccessTokenExpiring = 'addAccessTokenExpiring',
  SilentRenewError = 'AddSilentRenewError',
  UserLoaded = 'addUserLoaded',
  UserUnloaded = 'addUserUnloaded',
  UserSignedOut = 'addUserSignedOut',
  UserSessionChanged = 'addUserSessionChanged',
}